+++ web-ui/src/RuntimeLayoutOverlay.tsx
@@ -91,7 +91,7 @@ export default function RuntimeLayoutOverlay({
     for (const element of layout.getAllElements()) {
       if (isTurnoutElement(element) && element.outputMode === "accessory" && element.turnoutAddress === data.address) {
         element.turnoutClosed = data.closed;
-      } else if (element instanceof TrackTurnoutDoubleElementView) {
+      } else if (element instanceof TrackTurnoutDoubleElementView && element.outputMode === "accessory") {
         if (element.turnout1Address === data.address) element.turnout1Closed = data.closed;
         if (element.turnout2Address === data.address) element.turnout2Closed = data.closed;
       }
@@ -117,6 +117,23 @@ export default function RuntimeLayoutOverlay({
         element.barrierClosed = data.active === element.basicAccessoryClosedValue;
       }
     }
+    invalidate();
+  }), [layout, invalidate]);
+
+  useEffect(() => wsClient.on("vpinChanged", data => {
+    for (const element of layout.getAllElements()) {
+      if (isTurnoutElement(element) && element.outputMode === "vpin" && element.turnoutAddress === data.vpin) {
+        element.turnoutClosed = data.active;
+      } else if (element instanceof TrackTurnoutDoubleElementView && element.outputMode === "vpin") {
+        if (element.turnout1Address === data.vpin) element.turnout1Closed = data.active;
+        if (element.turnout2Address === data.vpin) element.turnout2Closed = data.active;
+      } else if (element instanceof TrackSignalElementView && element.outputMode === "vpin" && element.address <= data.vpin && element.lastAddress >= data.vpin) {
+        element.setValue(data.vpin, data.active);
+      } else if (element instanceof ButtonElementView && element.outputMode === "vpin" && element.address === data.vpin) {
+        element.on = data.active === element.activeValue;
+      }
+    }
+    layout.checkRoutes();
     invalidate();
   }), [layout, invalidate]);